requirements.txt
django
python-dotenv
pillow
mistralai
python-telegram-bot
django-jazzmin
celery
redis
django-celery-results
django-celery-beat